+++ src/main/kotlin/yandex/cloud/toolkit/configuration/function/deploy/DeployFunctionConfiguration.kt
@@ -11,8 +11,11 @@ import com.intellij.openapi.application.invokeLater
 import com.intellij.openapi.components.BaseState
 import com.intellij.openapi.options.SettingsEditor
 import com.intellij.openapi.project.Project
+import com.intellij.util.text.nullize
 import yandex.cloud.api.serverless.functions.v1.FunctionOuterClass
 import yandex.cloud.toolkit.api.resource.impl.model.CloudFunctionVersion
+import yandex.cloud.toolkit.api.resource.impl.model.CloudServiceAccount
+import yandex.cloud.toolkit.api.resource.impl.model.VPCNetwork
 import yandex.cloud.toolkit.process.FunctionDeployProcess
 import yandex.cloud.toolkit.process.RunContentController
 import yandex.cloud.toolkit.util.logger
@@ -25,7 +28,7 @@ class DeployFunctionConfiguration(name: String?, factory: ConfigurationFactory, 
     }
 
     override fun getConfigurationEditor(): SettingsEditor<out RunConfiguration> =
-        DeployFunctionConfigurationEditor(project, null, null, null, null)
+        DeployFunctionConfigurationEditor(project, null, FunctionDeployResources())
 
     override fun getState(executor: Executor, environment: ExecutionEnvironment): RunProfileState {
         if (state.functionId.isNullOrEmpty()) {
@@ -77,6 +80,11 @@ class FunctionDeploySpec : BaseState() {
     var serviceAccountId by string()
     var envVariables by map<String, String>()
     var tags by list<String>()
+    var networkId by string()
+    var useSubnets by property(false)
+    var subnets by list<String>()
+
+    fun hasConnectivity(): Boolean = if (useSubnets) subnets.isNotEmpty() else !networkId.isNullOrEmpty()
 
     companion object {
 
@@ -98,6 +106,18 @@ class FunctionDeploySpec : BaseState() {
             memoryBytes = template.resources.memory
             serviceAccountId = template.serviceAccountId
             tags = template.tagsList.filterTo(mutableListOf()) { it != CloudFunctionVersion.LATEST_TAG }
+            networkId = template.connectivity.networkId.nullize()
+            if (template.connectivity.subnetIdCount > 0) {
+                useSubnets = networkId.isNullOrEmpty()
+                subnets = template.connectivity.subnetIdList
+            }
         }
     }
 }
+
+data class FunctionDeployResources(
+    val versions: List<CloudFunctionVersion>? = null,
+    val serviceAccounts: List<CloudServiceAccount>? = null,
+    val networks: List<VPCNetwork>? = null,
+    val runtimes: List<String>? = null,
+)