+++ src/main/kotlin/yandex/cloud/toolkit/configuration/function/deploy/DeployFunctionConfigurationEditor.kt
@@ -3,25 +3,24 @@ package yandex.cloud.toolkit.configuration.function.deploy
 import com.intellij.openapi.options.SettingsEditor
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.ValidationInfo
+import com.intellij.ui.ClickListener
 import com.intellij.ui.components.JBLabel
 import com.intellij.ui.components.JBTabbedPane
 import com.intellij.ui.components.fields.IntegerField
 import com.intellij.util.text.nullize
 import com.intellij.util.ui.JBUI
 import yandex.cloud.toolkit.api.resource.impl.model.CloudFunction
-import yandex.cloud.toolkit.api.resource.impl.model.CloudFunctionVersion
-import yandex.cloud.toolkit.api.resource.impl.model.CloudServiceAccount
 import yandex.cloud.toolkit.ui.component.*
 import yandex.cloud.toolkit.util.*
+import java.awt.BorderLayout
+import java.awt.event.MouseEvent
 import javax.swing.JComponent
 import javax.swing.JTextField
 
 class DeployFunctionConfigurationEditor(
     val project: Project,
     target: CloudFunction?,
-    versions: List<CloudFunctionVersion>?,
-    serviceAccounts: List<CloudServiceAccount>?,
-    runtimes: List<String>?,
+    val resources: FunctionDeployResources
 ) : SettingsEditor<DeployFunctionConfiguration>() {
 
     companion object {
@@ -35,20 +34,40 @@ class DeployFunctionConfigurationEditor(
         private val KEY_MEMORY_VALUES = intArrayOf(MIN_MEMORY, 512, 1024, 1536, MAX_MEMORY)
     }
 
-    private val functionField = CloudResourceField(project, target, CloudFunction.Descriptor.icon)
+    private val functionField = CloudResourceLabel(project, target, CloudFunction.Descriptor.icon)
+    private val tabs = JBTabbedPane()
 
-    private val runtimeField = FunctionRuntimeField(project, runtimes)
+    private val runtimeField = FunctionRuntimeField(project, resources.runtimes)
     private val entryPointField = JTextField()
     private val sourceFilesList = SourceFilesList(project)
     private val sourceFolderPolicyBox = SourceFolderPolicyBox()
     private val descriptionArea = LimitedTextArea("Description", MAX_DESCRIPTION_LENGTH)
 
-    private val serviceAccountField = ServiceAccountField(project, target?.group?.folder?.id, serviceAccounts)
+    private val serviceAccountField = ServiceAccountField(project, target?.group?.folder?.id, resources.serviceAccounts)
     private val timeoutField = IntegerField(null, 0, Integer.MAX_VALUE)
     private val memoryField = MemoryField(MIN_MEMORY, MAX_MEMORY, MEMORY_PART, DEFAULT_MEMORY, KEY_MEMORY_VALUES)
 
     private val envVariablesList = EnvironmentVariablesList()
-    private val tagsList = FunctionVersionTagsList(null, versions ?: emptyList())
+    private val tagsList = FunctionVersionTagsList(null, resources.versions ?: emptyList())
+
+    private val networkField = VPCNetworkField(project, target?.group?.folder?.id, resources.networks)
+    private val subnetsList = SpoilerPanel("Enter Subnets", VPCSubnetsList(project, target?.group?.folder?.id))
+
+    private var vpsVisible = false
+    private val vpcTab = YCUI.borderPanel()
+
+    private fun checkVPCAvailable(s: DeployFunctionConfiguration) {
+        val showVPC = s.state.hasConnectivity() || resources.versions?.any { it.data.hasConnectivity() } == true
+
+        if (vpsVisible != showVPC) {
+            if (showVPC) {
+                tabs.addTab("VPC", vpcTab)
+            } else {
+                tabs.removeTabAt(tabs.tabCount - 1)
+            }
+            vpsVisible = showVPC
+        }
+    }
 
     override fun resetEditorFrom(s: DeployFunctionConfiguration) {
         functionField.value = s.state.functionId ?: ""
@@ -62,10 +81,10 @@ class DeployFunctionConfigurationEditor(
         serviceAccountField.value = s.state.serviceAccountId ?: ""
         envVariablesList.entries = s.state.envVariables
         tagsList.tags = s.state.tags.toSet()
-    }
-
-    override fun disposeEditor() {
-        runtimeField.dispose()
+        networkField.value = s.state.networkId ?: ""
+        subnetsList.content.subnets = s.state.subnets
+        subnetsList.isOpened = s.state.useSubnets
+        checkVPCAvailable(s)
     }
 
     override fun applyEditorTo(s: DeployFunctionConfiguration) {
@@ -81,6 +100,9 @@ class DeployFunctionConfigurationEditor(
             serviceAccountId = serviceAccountField.value.nullize()
             envVariables = envVariablesList.entries.toMutableMap()
             tags = tagsList.tags.toMutableList()
+            networkId = networkField.value.nullize()
+            subnets = subnetsList.content.subnets.toMutableList()
+            useSubnets = subnetsList.isOpened
         })
     }
 
@@ -93,46 +115,75 @@ class DeployFunctionConfigurationEditor(
         else -> descriptionArea.checkRestrictions()
     }
 
-    override fun createEditor(): JComponent = YCUI.gridPanel {
-        YCUI.gridBag(horizontal = true){
-            JBLabel("Function ID ") addAs nextln(0.0)
-            functionField addAs next(1.0)
-
-            JBTabbedPane().apply {
-                border = JBUI.Borders.emptyTop(5)
-
-                addTab("Main", YCUI.gridPanel {
-                    YCUI.gridBag(horizontal = true) {
-                        JBLabel("Runtime ") addAs nextln(0.0)
-                        runtimeField addAs next(1.0)
-                        JBLabel("Entry Point ") addAs nextln(0.0)
-                        entryPointField addAs next(1.0)
-
-                        sourceFilesList addAs fullLine()
-                        JBLabel("Source Folders ") addAs nextln(0.0)
-                        sourceFolderPolicyBox addAs next(1.0)
-                    }
-                })
-
-                addTab("Parameters", YCUI.gridPanel {
-                    YCUI.gridBag(horizontal = true) {
-                        JBLabel("Service Account ID ") addAs nextln(0.0)
-                        serviceAccountField addAs next(1.0)
-                        JBLabel("Timeout (sec)") addAs nextln(0.0)
-                        timeoutField addAs next(1.0)
-                        JBLabel("Memory") addAs nextln(0.0).insetTop(3)
-                        memoryField addAs next().coverLine()
-                        envVariablesList addAs fullLine()
-                    }
-                })
-
-                addTab("Misc", YCUI.gridPanel {
-                    YCUI.gridBag(horizontal = true) {
-                        descriptionArea.withPreferredHeight(100) addAs fullLine()
-                        tagsList.withPreferredHeight(100)  addAs fullLine()
-                    }
-                })
-            } addAs fullLine()
+    private fun updateNetworkField() {
+        networkField.isEnabled = !subnetsList.isOpened
+        networkField.isFocusable = networkField.isEnabled
+        if (networkField.isEnabled) {
+            networkField.field.grabFocus()
+            networkField.field.setTextToTriggerEmptyTextStatus("")
+            networkField.field.emptyText.text = ""
+        } else {
+            subnetsList.content.grabFocus()
+            networkField.field.setTextToTriggerEmptyTextStatus(networkField.field.text)
+            networkField.field.emptyText.text = "Subnets mode selected (click to change)"
         }
+    }
+
+    override fun createEditor(): JComponent = YCUI.borderPanel {
+        subnetsList.addSpoilerListener(SpoilerToggleListener(this@DeployFunctionConfigurationEditor::updateNetworkField))
+        object : ClickListener() {
+            override fun onClick(event: MouseEvent, clickCount: Int): Boolean {
+                if (event.button == MouseEvent.BUTTON1) {
+                    if (!networkField.isEnabled) subnetsList.close()
+                    return true
+                }
+                return false
+            }
+        }.installOn(networkField.field)
+        updateNetworkField()
+
+        functionField.labeled("Function ID") addAs BorderLayout.NORTH
+
+        tabs.apply {
+            border = JBUI.Borders.emptyTop(5)
+
+            addTab("Main", YCUI.borderPanel {
+                YCUI.gridPanel {
+                    YCUI.gridBag(horizontal = true) {
+                        JBLabel("Runtime: ") addAs nextln(0.0)
+                        runtimeField addAs next(1.0)
+                        JBLabel("Entry Point: ") addAs nextln(0.0)
+                        entryPointField addAs next(1.0)
+                    }
+                } addAs BorderLayout.NORTH
+
+                sourceFilesList addAs BorderLayout.CENTER
+                sourceFolderPolicyBox.labeled("Source Folders") addAs BorderLayout.SOUTH
+            })
+
+            addTab("Parameters", YCUI.borderPanel {
+                YCUI.gridPanel {
+                    YCUI.gridBag(horizontal = true) {
+                        JBLabel("Service Account ID: ") addAs nextln(0.0)
+                        serviceAccountField addAs next(1.0)
+                        JBLabel("Timeout (sec): ") addAs nextln(0.0)
+                        timeoutField addAs next(1.0)
+                        JBLabel("Memory: ") addAs nextln(0.0).insetTop(3)
+                        memoryField addAs next().coverLine()
+                    }
+                } addAs BorderLayout.NORTH
+                envVariablesList addAs BorderLayout.CENTER
+            })
+
+            addTab("Misc", YCUI.borderPanel {
+                descriptionArea.withPreferredHeight(100) addAs BorderLayout.NORTH
+                tagsList.withPreferredHeight(100) addAs BorderLayout.CENTER
+            })
+
+            vpcTab.apply {
+                networkField.labeled("Network ID") addAs BorderLayout.NORTH
+                subnetsList.withPreferredHeight(120) addAs BorderLayout.CENTER
+            }
+        } addAs BorderLayout.CENTER
     }
 }