+++ src/main/kotlin/yandex/cloud/toolkit/ui/dialog/FunctionDeployDialog.kt
@@ -11,11 +11,10 @@ import com.intellij.openapi.ui.DialogWrapper
 import com.intellij.openapi.ui.ValidationInfo
 import yandex.cloud.toolkit.api.auth.CloudAuthData
 import yandex.cloud.toolkit.api.resource.impl.model.CloudFunction
-import yandex.cloud.toolkit.api.resource.impl.model.CloudFunctionVersion
-import yandex.cloud.toolkit.api.resource.impl.model.CloudServiceAccount
 import yandex.cloud.toolkit.api.resource.impl.model.latestVersion
 import yandex.cloud.toolkit.configuration.function.deploy.DeployFunctionConfiguration
 import yandex.cloud.toolkit.configuration.function.deploy.DeployFunctionConfigurationEditor
+import yandex.cloud.toolkit.configuration.function.deploy.FunctionDeployResources
 import yandex.cloud.toolkit.util.disposeWith
 import yandex.cloud.toolkit.util.text
 import yandex.cloud.toolkit.util.withPreferredWidth
@@ -27,18 +26,16 @@ class FunctionDeployDialog(
     val project: Project,
     val authData: CloudAuthData,
     val function: CloudFunction,
-    val versions: List<CloudFunctionVersion>,
-    val serviceAccounts: List<CloudServiceAccount>,
-    runtimes: List<String>,
+    val resources: FunctionDeployResources,
     templateConfiguration: DeployFunctionConfiguration?,
     useTemplateTags: Boolean,
 ) : DialogWrapper(true) {
 
-    private val editor = DeployFunctionConfigurationEditor(project, function, versions, serviceAccounts, runtimes)
+    private val editor = DeployFunctionConfigurationEditor(project, function, resources)
     private val saveConfigurationAction = SaveConfigurationAction()
 
     private val configuration: DeployFunctionConfiguration = templateConfiguration?.clone()
-        ?: DeployFunctionConfiguration.createTemplateConfiguration(project, versions.latestVersion?.data)
+        ?: DeployFunctionConfiguration.createTemplateConfiguration(project, resources.versions?.latestVersion?.data)
 
     init {
         if (!useTemplateTags) configuration.state.tags.clear()
@@ -46,15 +43,19 @@ class FunctionDeployDialog(
 
         okAction.text = "Deploy"
 
-        editor.resetFrom(configuration)
         editor.disposeWith(myDisposable)
+
         init()
         title = "Deploy Yandex.Cloud Function"
     }
 
     override fun doValidate(): ValidationInfo? = editor.doValidate()
 
-    override fun createCenterPanel(): JComponent = editor.component.withPreferredWidth(700)
+    override fun createCenterPanel(): JComponent {
+        val panel = editor.component.withPreferredWidth(700)
+        editor.resetFrom(configuration)
+        return panel
+    }
 
     override fun doOKAction() {
         if (!okAction.isEnabled) return