+++ src/main/kotlin/yandex/cloud/toolkit/util/remote/resource/RemoteResource.kt
@@ -19,7 +19,6 @@ class RemoteResource<E>() {
         state = LoadedResourceState(value)
     }
 
-
     constructor(error: ResourceLoadingError) : this() {
         state = FailedResourceState(null, error)
     }